fix: setup of halogen, incandescent, and motors

+++ app/src/main/java/com/gemini/energy/service/device/lighting/Incandescent.kt
@@ -76,9 +76,9 @@ class Incandescent(computable: Computable<*>, utilityRateGas: UtilityRate, utili
             val config = lightingConfig(ELightingType.Incandescent)
             percentPowerReduced = config[ELightingIndex.PercentPowerReduced.value] as Double
 
-            peakHours = featureData["Peak Hours"]!! as Double
-            partPeakHours = featureData["Part Peak Hours"]!! as Double
-            offPeakHours = featureData["Off Peak Hours"]!! as Double
+            peakHours = (featureData["Peak Hours"]!! as Int).toDouble()
+            partPeakHours = (featureData["Part Peak Hours"]!! as Int).toDouble()
+            offPeakHours = (featureData["Off Peak Hours"]!! as Int).toDouble()
 
             alternateActualWatts = featureData["Alternate Actual Watts"]!! as Double
             alternateNumberOfFixtures = featureData["Alternate Number of Fixtures"]!! as Int
